requirements.txt
fastapi
fastapi-cache2[redis]
redis